Change the way configmap is reconciled
Instead of working by diffs, we always collect all the CRs from the
client (no impacts on the apiserver since they are cached), build the
configmap, check if there are any difference and then update (or create,
if the configmap did not exist in first place).

This makes it easier to handle corner cases and to add more CRDs in the
future.

As an effect of this change, the configmap is not removed anymore but
the configuration is cleared, since MetalLB ignores deleted configmaps.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>