Move the CRD controllers to a single controller
Having multiple controllers contributing to the same configmap is racy.
We had an occurrence in CI where both bfd profile and addresspool were
received, and the first write was slower than the second, resulting in a
malformed configmap.

Here, we reduce everything to a single controller watching for all the
resources, making the logic being mono-goroutine and removing the risk
for races.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>