Bundle manifests: remove unused volume and volumeMount
Removing the volume and volumeMount since OLM creates and
mounts a set of certs.

The volume and volumeMount are needed for certmanager when
deploying the operator through manifests.

Doing this to avoid confusion and creation of unused objects
when deploying with OLM.

The patch is taken from the memcached-operator sample project:
https://github.com/operator-framework/operator-sdk/blob/dad3e83dc7dabf0a7bea4eb5d4b52ea0911e8943/testdata/go/v3/memcached-operator/config/manifests/kustomization.yaml#L12