Refactor metallb-operator e2e tests (#135)
Created helper functions for repetitive code:
AssertCOnfigMapMatchYAML: this function gets a string and asserts
that the current configmap matches that. this functionality
repeats a lot in the code.

IsValidationWebhookRunning and IsValidationWebhookEnabled -
both logic was written for BGPPeer and AddressPools, and
if a new validation webhook may be added, the code will be duplicated
again. so I extracted it to functions.