E2E add AfterEach delete CRs
In the outer AfterEach of the e2e added a call
for a function that deletes all the CRs that were
created, and validates that they were removed completely.

usually the tests will clean those by themselves but in case
of test failure this func will prevent resource from remaining
in the cluster.

Also, added the prefix "test-" to cr's metadata name in case
we will want in the future to add the option to run the e2e
while the user is using the metallb operator.
i.e. we could delete them specifically.

Fix #123