Set the entry points in ocp, clean controller's security context
The values are or-ed, so we can't just override them. We clean and
override the security context of the controller pod when the object is
created.

Also, we set the entry points of the two containers in case of OCP
(where there's one single image), and we always disable the controller's
webhook mode in the operator as it's deployed externally as a separate
pod.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>