E2E tests: Avoid flakes: Check pods status in an Eventually
The operator pods can move from one node to another
in case of a node reboot. If we are checking a pod
status just at the same moment it's moved from one
node to another the test will fail. To avoid this flake
we should do the pods validation in an Eventually.