Add Periodic MetalLB Bump workflow
This commit adds a new workflow to periodically align
with the upstream MetalLB repository.

This is setup to run daily at 7AM and 7PM (GMT+2 clock),
run the make `bump_metallb` and if differences exist,
then open a PR with the bump,
and on the commit message added details of the old commit,
and the new commit, just so it will be easier
for developers to maintain and keep track of changes.

Created metallb_ref.txt as single source of truth for
the metallb commit hash.

Also deleted the `make test` from the bump_metallb, because we want
the workflow not to fail and to open PR even if the unit tests are
failing, so we could catch and fix that.

Notice: Every time this workflow runs it fetches latest metallb
and force-push to already open pr under the "bumpmetallb" branch.

Signed-off-by: liornoy <lnoy@redhat.com>