Fix the generate-metallb-manifests.sh script
In metallb we merged a commit (cb3bcb8) that separates the
config/crd/crd-conversion-patch.yaml into two different yamls
under the new config/crd/patches directory.

This change breaks the generate-metallb-manifests script that
fails to copy the old yaml.

This commit not only corrects the script itself, but also
aligns with metallb and seperates the crd-conversion-patch.yaml
into /patches dir, and updates the kustomization.yaml to reflect that.

Signed-off-by: liornoy <lnoy@redhat.com>