Move the k8sreporter dump into FailHandler
This commit moves the reporter.Dump() call into the FailHandler.
We do this because previously, having the reporter dump in the
ReportAfterEach, we didn't get the desired logs because
the test namesapce would be cleaned beforehand in the AfterEach/
AfterAll phase. Also wee modify the skipCaller to point to the correct
line.

Signed-off-by: Lior Noy <lnoy@redhat.com>