Add frr-k8s manifests
We modify the generation script to fetch the frr-k8s manifests and run it.
We also change the path for MetalLB's helm chart so that the two charts are
contained in `bindata/deployment/helm` separately and align the current code
to match the new path.
This also bumps the metallb_commit to one that contains the relevant changes
in MetalLB

Signed-off-by: Ori Braunshtein <obraunsh@redhat.com>